+++ src/js/basket.js
@@ -1,5 +1,5 @@
 import {refs} from "./refs";
-export {userBasket, onClickBasketBackdrop, onEscKeyPressBasket, ESC_KEY_CODE}
+export {userBasket, onClickBasketBackdrop, onEscKeyPressBasket, ESC_KEY_CODE, onBasketShow}
 import userEventApi from "./api"
 
 const ESC_KEY_CODE = "Escape";
@@ -15,10 +15,26 @@ function onClickVipBtn (event) {
 
 refs.basketHead.addEventListener("click", onClickBasketHead);
 function onClickBasketHead (event) {
-    refs.basketModal.classList.toggle("hidden")
+    onBasketShow()
+}
+function onBasketShow() {
+    refs.basketModal.classList.toggle("hidden");
+    refs.basketQuantity.textContent = userBasket.totalQuantity;
+    refs.basketNum.textContent = userBasket.totalQuantity;
+    if (userBasket.totalQuantity !== 0 && refs.basketContainer.classList.contains("hidden")) refs.basketContainer.classList.remove("hidden")
+    renderBasketMarkup(userBasket.contentShoppingCart)/// Данные с именем события
     refs.basketBackdrop.addEventListener("click", onClickBasketBackdrop)
     window.addEventListener("keydown", onEscKeyPressBasket);
 }
+function renderBasketMarkup(data) {
+    
+    refs.basketMarkupContainer.innerHTML = '';
+    let markup = '';
+    data.forEach(name => {
+      markup += `<li><div><div class="modal-basket__name">${name}</div></div></li>`;
+    });
+    refs.basketMarkupContainer.insertAdjacentHTML('beforeend', markup);
+  }
 function onClickBasketBackdrop(event) {
     if(event.currentTarget === event.target) {
         refs.basketModal.classList.toggle("hidden")
@@ -99,19 +115,19 @@ class Basket {
 
 let userBasket = {};
 function firstLoadPage () {
-    
     if (!localStorage.getItem("userBasket")) return userBasket = new Basket; //должно создаваться при загрузке Фетча
     const oldUserBasket = (JSON.parse(localStorage.getItem("userBasket")))
     userBasket = new Basket
     return Object.assign(userBasket, oldUserBasket)
-  
 }
+
 firstLoadPage()
 
 if (userBasket.totalQuantity !== 0 && refs.basketContainerHead.classList.contains("hidden")) {
     refs.basketContainerHead.classList.remove("hidden")
     refs.basketNumHead.textContent = userBasket.totalQuantity;
 }
+
